fixed all bugs but recognized problem in the logic
All the buttons work now but I realized that this exact approach is not
going to work. The fact that the RODs are associated with specific FECs
means that if you renormalize the FECs, which is done by changing the
calib structure and then reprocessing the video, the RODs will get
reapplied wrong when the video is reprocessed.