corrected ROD application logic
Made code more flexible so that RODs will be applied to the eyelidtrace
values that were in place at the time of ROD generation (e.g., they will
always go to the same frames instead of going to different frames when
the FEC measurement changes as a result of ROD application). Does
require the user to not change the threshold after setting the RODs
though.