New branch to change how the calibration scale and offset are determined
I ran into a file where the eyelidpos exceeded 1 by a lot even though I was resetting the calib.scale. Want to change the logic so that resetting calib.scale allows the user to truly determine which frame = 1 FEC.

First step is to stop eyetrace from being forced to 1 to allow better troubleshooting.